+++ rules/get_chirps.py
@@ -1,16 +1,10 @@
 rule get_chirps:
-    input:
-        "data/{area}.shp",
     output:
-        "results/data/chirps_{area}.nc"
+        # temp("results/data/chirps.nc")
+        "results/data/chirps.nc"
     log:
-        "results/logs/get_chirps_{area}.log"
+        "results/logs/get_chirps.log"
     benchmark:
-        "results/benchmarks/get_chirps_{area}.benchmark.txt"
-    threads: 1
-    resources:
-        mem_mb=1000
-    conda:
-        "../envs/xarray.yml"
-    script:
-      "../scripts/get_chirps.py"
+        "results/benchmarks/get_chirps.benchmark.txt"
+    shell:
+      "wget 'https://data.chc.ucsb.edu/products/CHIRPS-2.0/global_monthly/netcdf/chirps-v2.0.monthly.nc' -o {output}"