+++ NEWS.md
@@ -1,1 +1,2 @@
+-   ideas 0.1.0.9002: get chirps, chelsa, tmf, modis
 -   ideas 0.1.0.9001: init